rename glyphinfo to glyphs
